+++ KVASS/Properties/AssemblyVersion.cs
@@ -4,4 +4,4 @@
   
  using System.Reflection;
 
- [assembly: AssemblyVersion("0.8.0.0")]
+ [assembly: AssemblyVersion("0.8.1.0")]