+++ ChangeLog.md
@@ -1,5 +1,9 @@
 ChangeLog:
 
+## Version 0.8.1
+ * fixed significantDigits in the message
+ * renamed folder to the KVASS
+
 ## Version 0.8.0
  * clearing code